refactor(compliance): split code
